PostgreSQL: rename rust directory
